Updates to move logic
Fixed a couple bugs where player would check columns that were full or would try to move in columns that were full. Currently, player does not see B_B as a potential sequence of two. I would like the logic to see one piece on each side and consider it a potential sequence of 2. Once that is done, I would like to disregard potential sequences that can never reach the sequence length. For example |_B_R... where | is the wall. This sequence can never reach a sequence length of four, so we can treat it with lower priority.